+++ src/components/timelineComp/TimelineComp.tsx
@@ -11,7 +11,8 @@ type SliderPropeTypes = {
 }
 
 const Slider: FC<SliderPropeTypes> = ({ currentBlockNumber, saleInfo, config, constants }) => {
-  if (!saleInfo.saleStart || !saleInfo.endPrice) return null
+  if (!saleInfo.saleStart || (!saleInfo.endPrice && !saleInfo.price)) return null
+  const price = saleInfo.endPrice ? saleInfo.endPrice : saleInfo.price
   const saleDuration = (config.regionLength * constants.timeslicePeriod) / 2
 
   // Calculate percentages for each period
@@ -61,9 +62,7 @@ const Slider: FC<SliderPropeTypes> = ({ currentBlockNumber, saleInfo, config, co
         className="absolute top-0 -mt-1"
         style={{ left: `${(safeLeadinPercentage + 100) / 2}%` }}
       >
-        <p className="text-sm text-left mt-5 -ml-10">
-          Stable price = {Number(saleInfo.endPrice) / 10 ** 12}
-        </p>
+        <p className="text-sm text-left mt-5 -ml-10">Stable price = {Number(price) / 10 ** 12}</p>
       </div>
       {/* Marker for Purchase Period */}
       <div className="absolute top-0 -mt-1" style={{ left: '100%' }}>
